feat(core): extract tracing subsystem into core/tracing/ (Phase F)
Move trace-context, trace-store, otel-exporter, otlp-transformer, and
otlp-protobuf (+ tests) from services/ to backend/src/core/tracing/.
Update request-tracing plugin to import from ../tracing/ (within core).
Create barrel index.ts and re-export shims at old services/ paths.
Resolves the core <-> observability circular dependency.

Part of #706
