+++ backend/src/core/tracing/index.ts
@@ -1,0 +1,6 @@
+// Barrel export for core/tracing
+export * from './trace-context.js';
+export * from './trace-store.js';
+export * from './otel-exporter.js';
+export * from './otlp-transformer.js';
+export { decodeOtlpProtobuf } from './otlp-protobuf.js';
